polish: styling, mechanics, and action wiring improvements
- Fix char limit off-by-one (>= 3 → > 3, allowing exactly 3)
- Fix knowledge decay rate (0.250 → 0.010 per tick)
- Add food consumption from inventory to survival system
- Add POST /auth/logout endpoint
- Add territory claim on settlement founding
- Add view fade-in animation, loading spinner, skeleton, button disabled states
- Add action parameter dialogs (group, settlement, structure, craft, teach)
- Wire frontend to backend APIs for group/settlement/structure creation
- Fix build_structure ::numeric cast for strict PG compatibility

+++ src/simulation/systems/playerActions.ts
@@ -226,7 +226,7 @@ async function processBuildStructure(
   const newProgress = Math.min(100, Number(structure.construction_progress) + laborAmount);
 
   await client.query(
-    `UPDATE structures SET construction_progress = $1, completed_at = CASE WHEN $1 >= 100 AND completed_at IS NULL THEN now() ELSE completed_at END WHERE id = $2`,
+      `UPDATE structures SET construction_progress = $1, completed_at = CASE WHEN $1::numeric >= 100 AND completed_at IS NULL THEN now() ELSE completed_at END WHERE id = $2`,
     [newProgress, structureId]
   );
